% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{appendArg}
\alias{appendArg}
\title{Append an argument on a list}
\usage{
appendArg(args, argName, argValue, forced)
}
\arguments{
\item{args}{The original argument list.}

\item{argName}{The name of the argument to add.}

\item{argValue}{The value of the argument to add.}

\item{forced}{a \code{logical} value indicating if the argument with the same
name already existed, whether it should be added forcedly.}
}
\value{
An argument list.
}
\description{
Append an argument on a list of arguments
}
