% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backprop_surv}
\alias{backprop_surv}
\title{Backprop Surv (Internal)}
\usage{
backprop_surv(
  n_hidden,
  w_ini,
  load_param,
  weight,
  bias,
  x,
  y,
  w,
  valid,
  x_valid,
  y_valid,
  w_valid,
  activ,
  n_epoch,
  n_batch,
  model_type,
  learning_rate,
  l1_reg,
  l2_reg,
  early_stop,
  early_stop_det,
  learning_rate_adaptive,
  rho,
  epsilon,
  beta1,
  beta2,
  loss_f
)
}
\arguments{
\item{n_hidden}{Hidden layer numbers.}

\item{w_ini}{Initial weights.}

\item{load_param}{Whether to load parameters.}

\item{weight}{Weights.}

\item{bias}{Biases.}

\item{x}{Training.}

\item{y}{Training.}

\item{w}{Training.}

\item{valid}{Whether to use a validation set.}

\item{x_valid}{Validation.}

\item{y_valid}{Validation.}

\item{w_valid}{Validation.}

\item{activ}{Activation function.}

\item{n_epoch}{Number of epochs.}

\item{n_batch}{Batch size.}

\item{model_type}{Model type.}

\item{learning_rate}{(Initial) learning rate.}

\item{l1_reg}{L1-penalty.}

\item{l2_reg}{L2-penalty.}

\item{early_stop}{Whether to early stop.}

\item{early_stop_det}{Number of epochs to determine early-stop.}

\item{learning_rate_adaptive}{Adaptive learning rate adjustment method.}

\item{rho}{Parameter.}

\item{epsilon}{Parameter.}

\item{beta1}{Parameter.}

\item{beta2}{Parameter.}

\item{loss_f}{Loss function.}
}
\value{
A list of outputs
}
\description{
Backprop Surv (Internal)
}
