% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-permfit.R
\name{cox_logl}
\alias{cox_logl}
\title{Log-likelihood for Cox-Model (Internal)}
\usage{
cox_logl(h, y, e)
}
\arguments{
\item{h}{Log hazard.}

\item{y}{Event time.}

\item{e}{Event status.}
}
\value{
Returns log-likelihood.
}
\description{
Log-likelihood for Cox-Model (Internal)
}
