\name{deepTL-package}
\alias{deepTL-package}
\alias{deepTL}
\docType{package}
\title{
\packageTitle{deepTL}
}
\description{
\packageDescription{deepTL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{deepTL}
\packageIndices{deepTL}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{deepTL}

Maintainer: \packageMaintainer{deepTL}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
