% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R, R/1-2-methods.R
\docType{class}
\name{dnnet-class}
\alias{dnnet-class}
\alias{show,dnnet-method}
\title{An S4 class containing a deep neural network}
\usage{
\S4method{show}{dnnet}(object)
}
\arguments{
\item{object}{A \code{dnnet} object.}
}
\description{
An S4 class containing a deep neural network
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{dnnet} object.
}}

\section{Slots}{

\describe{
\item{\code{norm}}{A list, containing the centers and s.d.'s of x matrix and y vector (if numeric)}

\item{\code{weight}}{A list of matrices, weight matrices in the fitted neural network}

\item{\code{bias}}{A list of vectors, bias vectors in the fitted neural network}

\item{\code{loss}}{The minimum loss acheived from the validate set}

\item{\code{label}}{If the model is classification, a character vectors containing class labels}

\item{\code{model.type}}{Either "Classification" or "Regression"}

\item{\code{model.spec}}{Other possible information}
}}

\seealso{
\code{\link{dnnetInput-class}}\cr
}
