% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{dnnetEnsemble-class}
\alias{dnnetEnsemble-class}
\title{An S4 class containing an ensemble of deep neural networks}
\description{
An S4 class containing an ensemble of deep neural networks
}
\section{Slots}{

\describe{
\item{\code{model.list}}{A list of dnnet models}

\item{\code{model.type}}{Either "Classification" or "Regression"}

\item{\code{model.spec}}{Other possible information}

\item{\code{loss}}{A numeric vector of loss from all DNNs}

\item{\code{keep}}{Whether the model is kept when put in the predict function}
}}

\seealso{
\code{\link{dnnet-class}}\cr
\code{\link{ensemble_dnnet}}\cr
}
