% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R, R/1-2-methods.R
\docType{class}
\name{dnnetInput-class}
\alias{dnnetInput-class}
\alias{show,dnnetInput-method}
\title{An S4 class containing predictors (x), response (y) and sample weights (w)}
\usage{
\S4method{show}{dnnetInput}(object)
}
\arguments{
\item{object}{A \code{dnnetInput} object.}
}
\description{
An S4 class containing predictors (x), response (y) and sample weights (w)
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{dnnetInput} object.
}}

\section{Slots}{

\describe{
\item{\code{x}}{A numeric matrix, the predictors}

\item{\code{y}}{A factor or numeric vector, either the class labels or continuous responses}

\item{\code{w}}{A numeric vector, sample weights}
}}

\seealso{
\code{\link{dnnet-class}}\cr
}
