% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{dnnetSurvInput-class}
\alias{dnnetSurvInput-class}
\title{An S4 class containing predictors (x), censoring time (y),
sample weights (w) and event status (e)}
\description{
An S4 class containing predictors (x), censoring time (y),
sample weights (w) and event status (e)
}
\section{Slots}{

\describe{
\item{\code{x}}{A numeric matrix, the predictors}

\item{\code{y}}{A numeric vector, censoring time}

\item{\code{w}}{A numeric vector, sample weights}

\item{\code{e}}{A numeric vector, event status}
}}

\seealso{
\code{\link{dnnet-class}}\cr
}
