% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-1-double-deep-tl.R
\name{double_deepTL}
\alias{double_deepTL}
\title{The algorithm for double deep treatment learning}
\usage{
double_deepTL(object, en_dnn_ctrl1 = NULL, en_dnn_ctrl2 = NULL,
  methods = c("revised-semi", "semi", "cov-adj", "ipw1", "ipw2", "ipw3",
  "double-robust"))
}
\arguments{
\item{object}{A \code{dnnetInput} or a \code{dnnetSurvInput} object, the training set.}

\item{en_dnn_ctrl1}{A list of parameters to be passed to \code{dnnet} for model 1.}

\item{en_dnn_ctrl2}{A list of parameters to be passed to \code{dnnet} for model 2.}

\item{methods}{Methods included in this analysis.}
}
\value{
Returns a \code{list} of results.
}
\description{
The algorithm for double deep treatment learning in comparative effectiveness analysis.
}
