% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-ensemble.R
\name{ensemble_dnnet}
\alias{ensemble_dnnet}
\title{An ensemble model of DNNs}
\usage{
ensemble_dnnet(
  object,
  n.ensemble = 100,
  esCtrl,
  unbalance.trt = c("None", "Over", "Under", "Balance")[1],
  bootstrap = TRUE,
  prop.train = 1,
  prop.keep = 1,
  best.opti = TRUE,
  min.keep = 10,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{dnnetInput} or a \code{dnnetSurvInput} object, the training set.}

\item{n.ensemble}{The number of DNNs to be fit}

\item{esCtrl}{A list of parameters to be passed to \code{dnnet}.}

\item{unbalance.trt}{Treatment for unbalanced labels.}

\item{bootstrap}{Indicator for whether a bootstrap sampling is used.}

\item{prop.train}{If bootstrap == FALSE, a training/validation cut for the data will be used (0, 1).}

\item{prop.keep}{The proportion of DNNs to be kept in the ensemble.}

\item{best.opti}{Whether to run the algorithm to keep the optimal subset of DNNs.}

\item{min.keep}{Minimal number of DNNs to be kept.}

\item{verbose}{Whether the progress to be printed.}
}
\value{
Returns a \code{dnnetEnsemble} object.
}
\description{
Fit a an ensemble model of DNNs
}
\seealso{
\code{\link{dnnetEnsemble-class}}\cr
\code{\link{dnnetInput-class}}\cr
}
