% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{getSplitDnnet}
\alias{getSplitDnnet}
\title{A function to generate indice}
\usage{
getSplitDnnet(split, n)
}
\arguments{
\item{split}{As in \code{\link{dnnetInput-class}}.}

\item{n}{Sample size}
}
\value{
Returns a integer vector of indice.
}
\description{
A function to generate indice
}
\seealso{
\code{\link{dnnetInput-class}}
}
