% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{importDnnet}
\alias{importDnnet}
\title{Import Data to create a \code{dnnetInput} object.}
\usage{
importDnnet(x, y, w = rep(1, length(y)))
}
\arguments{
\item{x}{A \code{matrix} containing all samples/variables. It has to be \code{numeric}
and cannot be left blank. Any variable with missing value will be removed.}

\item{y}{A \code{numeric} or \code{factor} vector, indicating a continuous outcome or class label.}

\item{w}{A \code{numeric} vector, the sample weight. Will be 1 if left blank.}
}
\value{
An \code{dnnetInput} object.
}
\description{
Import Data to create a \code{dnnetInput} object.
}
\seealso{
\code{\link{dnnetInput-class}}
}
