% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{importDnnetSurv}
\alias{importDnnetSurv}
\title{Import Data to create a \code{dnnetSurvInput} object.}
\usage{
importDnnetSurv(x, y, w = rep(1, length(y)), e)
}
\arguments{
\item{x}{A \code{matrix} containing all samples/variables. It has to be \code{numeric}
and cannot be left blank. Any variable with missing value will be removed.}

\item{y}{A \code{numeric} vector, indicating the censoring time.}

\item{w}{A \code{numeric} vector, the sample weight. Will be 1 if left blank.}

\item{e}{A \code{numeric} vector, the event status.}
}
\value{
An \code{dnnetInput} object.
}
\description{
Import Data to create a \code{dnnetSurvInput} object.
}
\seealso{
\code{\link{dnnetInput-class}}
}
