% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{importTrt}
\alias{importTrt}
\title{Import Data to create a \code{trtInput} object.}
\usage{
importTrt(x, y, z)
}
\arguments{
\item{x}{A \code{matrix} containing all samples/variables. It has to be \code{numeric}
and cannot be left blank. Any variable with missing value will be removed.}

\item{y}{A \code{numeric} vector, indicating the censoring time.}

\item{z}{The treatment assignment.}
}
\value{
An \code{trtInput} object.
}
\description{
Import Data to create a \code{trtInput} object.
}
\seealso{
\code{\link{trtInput-class}}
}
