% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{itrMod-class}
\alias{itrMod-class}
\title{An S4 class containing an individualized treatment rule (ITR) object}
\description{
An S4 class containing an individualized treatment rule (ITR) object
}
\section{Slots}{

\describe{
\item{\code{model}}{a model for ITR prediction}

\item{\code{framework}}{a character for ITR framework}
}}

\seealso{
\code{\link{dnnet-class}}\cr
}
