% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-permfit.R
\name{log_lik_diff}
\alias{log_lik_diff}
\title{Log-likelihood Difference (Internal)}
\usage{
log_lik_diff(model.type, y_hat, y_hat0, object)
}
\arguments{
\item{model.type}{Type of the model.}

\item{y_hat}{Y hat.}

\item{y_hat0}{Another Y hat.}

\item{object}{Data object.}
}
\value{
Returns log-likelihood difference.
}
\description{
Log-likelihood Difference (Internal)
}
