% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-permfit.R
\name{mod_permfit}
\alias{mod_permfit}
\title{Model passed to PermFIT (Internal)}
\usage{
mod_permfit(method, model.type, object, ...)
}
\arguments{
\item{method}{Name of the model.}

\item{model.type}{Type of model.}

\item{object}{A dnnetInput or dnnetSurvInput object.}

\item{...}{Orger parameters passed to the model.}
}
\value{
Returns a specific model.
}
\description{
Model passed to PermFIT (Internal)
}
