% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-ensemble.R
\name{neg_log_lik_}
\alias{neg_log_lik_}
\title{Negative log-likelihood (Internal)}
\usage{
neg_log_lik_(model.type, y, y_hat, alpha = NULL)
}
\arguments{
\item{model.type}{Type of the model.v}

\item{y_hat}{Y hat.x}

\item{alpha}{Other distributional parameter.}
}
\value{
Returns log-likelihood
}
\description{
Negative log-likelihood (Internal)
}
