% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{PermFIT-class}
\alias{PermFIT-class}
\title{An S4 class containing an permutation-based feature importance test (PermFIT) object}
\description{
An S4 class containing an permutation-based feature importance test (PermFIT) object
}
\section{Slots}{

\describe{
\item{\code{model}}{a model for ITR prediction}

\item{\code{importance}}{a data.frame of importance}

\item{\code{block_importance}}{a data.frame of block importance}

\item{\code{validation_index}}{indices for cross-fitting}

\item{\code{y_hat}}{cross-fitted prediction or prediction from validation}
}}

