% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-permfit.R
\name{permfit}
\alias{permfit}
\title{PermFIT: A permutation-based feature importance test.}
\usage{
permfit(
  train,
  validate = NULL,
  k_fold = 5,
  n_perm = 100,
  pathway_list = list(),
  method = c("ensemble_dnnet", "random_forest", "lasso", "linear", "svm", "dnnet")[1],
  shuffle = NULL,
  ...
)
}
\arguments{
\item{train}{An dnnetInput or dnnetSurvInput object.}

\item{validate}{A validation dataset is required when k_fold = 0.}

\item{k_fold}{K-fold cross-fitting. If not, set k_fold to zero.}

\item{n_perm}{Number of permutations repeated.}

\item{pathway_list}{A list of pathways to be jointly tested.}

\item{method}{Models, including \code{ensemble_dnnet} for ensemble deep
neural networks, \code{random_forest} for random forests or random
survival forests, \code{lasso} for linear/logistic/cox lasso, \
\code{linear} for linear/logistic/coxph regression, \code{svm}
for svms with Gaussian kernels, and \code{dnnet} with single deep
neural network.}

\item{shuffle}{If shuffle is null, the data will be shuffled for
cross-fitting; if random shuffle is not desired, please provide
a bector of numbers for cross-fitting indices}

\item{...}{Additional parameters passed to each method.}
}
\value{
Returns a PrmFIT object.
}
\description{
PermFIT: A permutation-based feature importance test.
}
