% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1-predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,dnnet-method}
\alias{predict,dnnetEnsemble-method}
\title{Predict function for the package}
\usage{
\S4method{predict}{dnnet}(object, newData, type, ...)

\S4method{predict}{dnnetEnsemble}(object, newData, type, ...)
}
\arguments{
\item{object}{A \code{dnnet} or \code{dnnetEnsemble} object.}

\item{newData}{A matrix with the same number of columns in the input data.}

\item{type}{Consistent with model.type in the \code{object}.}

\item{...}{Other parameters.}
}
\value{
A numeric vector for regression or a matrix of probabilities for each class for classification.
}
\description{
The prediction function for \code{dnnet} or \code{dnnetEnsemble} object.
}
\seealso{
\code{\link{dnnet-class}}\cr
\code{\link{dnnetEnsemble-class}}\cr
}
