% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-permfit.R
\name{predict_mod_permfit}
\alias{predict_mod_permfit}
\title{Model prediction passed to PermFIT (Internal)}
\usage{
predict_mod_permfit(mod, object, method, model.type)
}
\arguments{
\item{mod}{Model for prediction.}

\item{object}{A dnnetInput or dnnetSurvInput object.}

\item{method}{Name of the model.}

\item{model.type}{Type of the model.}
}
\value{
Returns predictions.
}
\description{
Model prediction passed to PermFIT (Internal)
}
