% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-control.R
\name{removeArg}
\alias{removeArg}
\title{Remove an argument on a list}
\usage{
removeArg(args, argName)
}
\arguments{
\item{args}{The original argument list.}

\item{argName}{The name of the argument to be remmoved. If the name does not exist,
no argument will be removed.}
}
\value{
An argument list.
}
\description{
Remove an argument on a list of arguments
}
