% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,dnnetInput-method}
\alias{show,dnnet-method}
\title{Method show for the package}
\usage{
\S4method{show}{dnnetInput}(object)

\S4method{show}{dnnet}(object)
}
\arguments{
\item{object}{A \code{dnnet} or \code{dnnetInput} object.}
}
\description{
The method show for \code{dnnetInput} or \code{dnnet} object.
}
\seealso{
\code{\link{dnnetInput-class}}\cr
\code{\link{dnnet-class}}\cr
}
