% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{splitDnnet}
\alias{splitDnnet}
\title{A function to split the \code{dnnetInput} object into a list of two \code{dnnetInput} objects:
one names train and the other named valid.}
\usage{
splitDnnet(object, split)
}
\arguments{
\item{object}{A \code{dnnetInput} object.}

\item{split}{A character, numeric variable or a numeric vector declaring a way to split
the \code{dnnetInput}. If it's number between 0 and 1, all samples will be split into two subsets
randomly, with the \code{train} containing such proportion of all samples and \code{valid} containing
the rest. If split is a character and is "bootstrap", the \code{train} will be a bootstrap sample
of the original data set and the \code{valid} will contain out-of-bag samples. If split is a vector
of integers, the \code{train} will contain samples whose indice are in the vector, and \code{valid} will
contain the rest.}
}
\value{
Returns a list of two \code{dnnetInput} objects.
}
\description{
A function to split the \code{dnnetInput} object into a list of two \code{dnnetInput} objects:
one names train and the other named valid.
}
\seealso{
\code{\link{dnnetInput-class}}
}
