% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-import.R
\name{splitTrt}
\alias{splitTrt}
\title{A function to split the \code{trtInput} object into a list of two \code{trtInput} objects:
one names train and the other named valid.}
\usage{
splitTrt(object, split)
}
\arguments{
\item{object}{A \code{trtInput} object.}

\item{split}{A character, numeric variable or a numeric vector declaring a way to split
the \code{trtInput}. If it's number between 0 and 1, all samples will be split into two subsets
randomly, with the \code{train} containing such proportion of all samples and \code{valid} containing
the rest. If split is a character and is "bootstrap", the \code{train} will be a bootstrap sample
of the original data set and the \code{valid} will contain out-of-bag samples. If split is a vector
of integers, the \code{train} will contain samples whose indice are in the vector, and \code{valid} will
contain the rest.}
}
\value{
Returns a list of two \code{trtInput} objects.
}
\description{
A function to split the \code{trtInput} object into a list of two \code{trtInput} objects:
one names train and the other named valid.
}
\seealso{
\code{\link{trtInput-class}}
}
