% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{trtInput-class}
\alias{trtInput-class}
\title{An S4 class containing predictors (x), response (y) and treatment assignment (z)}
\description{
An S4 class containing predictors (x), response (y) and treatment assignment (z)
}
\section{Slots}{

\describe{
\item{\code{x}}{A numeric matrix, the predictors}

\item{\code{y}}{A factor or numeric vector, either the class labels or continuous responses}

\item{\code{z}}{Treatment assignment}
}}

